requirements.txt
ruamel.yaml
netifaces
